requirements.txt
branca==0.3.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
folium==0.7.0
geographiclib==1.49
geopy==1.18.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
gunicorn==19.9.0
psycopg2-binary==2.7.7
